+++ core/app_shared.go
@@ -67,6 +67,10 @@ func (app *application) sharedGetAppConfig(appTypeIdentifier string, orgID *stri
 	return appConfigs, nil
 }
 
+func (app *application) sharedGetAppAssetFile(orgID string, appID string, name string) (*model.AppAsset, error) {
+	return app.storage.FindAppAsset(orgID, appID, name)
+}
+
 func (app *application) sharedGetAccount(cOrgID string, cAppID string, accountID string) (*model.Account, error) {
 	account, err := app.getAccount(nil, cOrgID, cAppID, accountID)
 	if err != nil {
@@ -132,10 +136,22 @@ func (app *application) sharedUpdateAccountUsername(accountID string, appID stri
 			return errors.ErrorData(logutils.StatusInvalid, model.TypeAccountUsername, logutils.StringArgs(username+" taken")).SetStatus(utils.ErrorStatusUsernameTaken)
 		}
 
-		//2. update the username
-		err = app.storage.UpdateAccountUsername(context, accountID, username)
+		account, err := app.getAccount(context, orgID, appID, accountID)
 		if err != nil {
-			return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccountUsername, nil, err)
+			return errors.WrapErrorAction(logutils.ActionGet, model.TypeAccount, &logutils.FieldArgs{"id": accountID}, err)
+		}
+
+		//2. save the username as a new account identifier or update an existing one
+		added, err := app.auth.AddAccountUsername(context, account, username)
+		if err != nil {
+			return errors.WrapErrorAction(logutils.ActionSave, model.TypeAccountUsername, nil, err)
+		}
+		if !added {
+			// the username could not be added as a new identifier, so try updating an existing username identifier
+			err = app.storage.UpdateAccountUsername(context, accountID, username)
+			if err != nil {
+				return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeAccountUsername, nil, err)
+			}
 		}
 
 		return nil