+++ CHANGELOG.md
@@ -7,16 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Handle client asset files [#436](https://github.com/rokwire/core-building-block/issues/436)
 - WebAuthn authentication [#659](https://github.com/rokwire/core-building-block/issues/659)
 - Searching follows looks for substring matches [#670](https://github.com/rokwire/core-building-block/issues/670)
 - Support following accounts [#667](https://github.com/rokwire/core-building-block/issues/667)
 - Device ID not nullable [#672](https://github.com/rokwire/core-building-block/issues/672)
 - Account secrets [#677](https://github.com/rokwire/core-building-block/issues/677)
 - Set up token duration policies for app/orgs [#472](https://github.com/rokwire/core-building-block/issues/472)
+- Expose account date created and updated through API
 ### Changed
 - Decouple authentication and verification mechanisms [#665](https://github.com/rokwire/core-building-block/issues/665)
 - Refactor account auth types [#674](https://github.com/rokwire/core-building-block/issues/674)
 - Improve request context logging
+- Upgrade dependencies
+- Separate auth claims checks from permissions checks
+### Fixed
+- Fix email and code sign up
+- Allow update username API to create username
 
 ## [1.38.1] - 2024-03-05
 ### Fixed